# perf: make `EvmAccountSelectorList.renderItem`  faster - cp-7.50.0