--- a/app/components/UI/EvmAccountSelectorList/EvmAccountSelectorList.tsx
+++ b/app/components/UI/EvmAccountSelectorList/EvmAccountSelectorList.tsx
@@ -560,7 +560,7 @@ const EvmAccountSelectorList = ({
       keyExtractor={getKeyExtractor}
       renderItem={renderItem}
       // Increasing number of items at initial render fixes scroll issue.
-      initialNumToRender={999}
+      initialNumToRender={accounts.length} // Using the optimal numbers seems ok here.
       testID={ACCOUNT_SELECTOR_LIST_TESTID}
       {...props}
     />