--- a/app/components/Views/confirmations/__mocks__/controllers/other-controllers-mock.ts
+++ b/app/components/Views/confirmations/__mocks__/controllers/other-controllers-mock.ts
@@ -26,6 +26,9 @@ export const accountsControllerMock = {
               address: accountMock,
               metadata: {
                 name: 'Account 1',
+                keyring: {
+                  type: 'HD Key Tree',
+                },
               },
             },
           },