--- a/app/components/hooks/useAccounts/useAccounts.ts
+++ b/app/components/hooks/useAccounts/useAccounts.ts
@@ -186,6 +186,8 @@ const useAccounts = ({
           caipAccountId: `${internalAccount.scopes[0]}:${internalAccount.address}`,
           scopes: internalAccount.scopes,
           isLoadingAccount: accountBalance.isLoadingAccount,
+          // Keep reference to the internal account to avoid unnecessary lookup.
+          internalAccount,
         };
         // Calculate height of the account item.
         yOffset += 78;