--- a/app/components/hooks/useAccounts/useAccounts.types.ts
+++ b/app/components/hooks/useAccounts/useAccounts.types.ts
@@ -4,7 +4,7 @@ import { KeyringTypes } from '@metamask/keyring-controller';
 // External dependencies.
 import { AvatarTokenProps } from '../../../component-library/components/Avatars/Avatar/variants/AvatarToken/AvatarToken.types';
 import { CaipAccountId, CaipChainId } from '@metamask/utils';
-import { InternalAccount } from '@metamask/keyring-internal-api';
+import { AccountId } from '@metamask/accounts-controller';
 
 /**
  * Asset information associated with the account, which includes both the fiat balance and owned tokens.
@@ -25,6 +25,10 @@ export interface Assets {
  */
 export interface Account {
   /**
+   * Account ID.
+   */
+  id: AccountId;
+  /**
    * Account name.
    */
   name: string;
@@ -32,10 +36,6 @@ export interface Account {
    * Account address.
    */
   address: string;
-  /**
-   * Internal account reference.
-   */
-  internalAccount: InternalAccount;
   /**
    * Asset information associated with the account, which includes both the fiat balance and owned tokens.
    */