--- a/app/components/hooks/useGetTotalFiatBalanceCrossChains.tsx
+++ b/app/components/hooks/useGetTotalFiatBalanceCrossChains.tsx
@@ -5,7 +5,7 @@ import {
   selectNetworkConfigurations,
 } from '../../selectors/networkController';
 import { selectAccountsByChainId } from '../../selectors/accountTrackerController';
-import { hexToBN, toChecksumHexAddress } from '@metamask/controller-utils';
+import { hexToBN } from '@metamask/controller-utils';
 import { TokensWithBalances } from './useGetFormattedTokensPerChain';
 import {
   selectCurrencyRates,
@@ -159,6 +159,7 @@ export const useGetTotalFiatBalanceCrossChains = (
     return aggregatedBalPerAccount;
   }, [
     accounts,
+    addressesByAccountId,
     formattedTokensWithBalancesPerChain,
     allNetworks,
     currencyRates,