… app/components/UI/Bridge/Views/BridgeView/BridgeView.test.tsx
… app/components/UI/EvmAccountSelectorList/__snapshots__/EvmAccountSelectorList.test.tsx.snap
… app/components/UI/WalletAccount/__snapshots__/WalletAccount.test.tsx.snap
… app/components/Views/AccountConnect/AccountConnectMultiSelector/__snapshots__/AccountConnectMultiSelector.test.tsx.snap
… app/components/Views/AccountPermissions/__snapshots__/AccountPermissions.test.tsx.snap
… app/components/Views/AccountSelector/AccountSelector.test.tsx
… app/components/Views/AccountSelector/__snapshots__/AccountSelector.test.tsx.snap
… app/components/Views/confirmations/components/confirm/confirm-component.test.tsx
… app/components/Views/confirmations/components/info-root/info-root.test.tsx
… app/components/Views/confirmations/components/info/personal-sign/personal-sign.test.tsx
… app/components/Views/confirmations/components/info/typed-sign-v1/typed-sign-v1.test.tsx
… app/components/Views/confirmations/components/info/typed-sign-v3v4/typed-sign-v3v4.test.tsx
… app/components/Views/confirmations/components/rows/account-network-info-row/account-network-info-row.test.tsx
… app/components/hooks/useAccounts/useAccounts.test.ts
… app/selectors/accountsController.test.ts
… app/util/test/accountsControllerTestUtils.ts
… app/util/test/confirm-data-helpers.ts
… app/util/test/keyringControllerTestUtils.ts