## **Description**

We were doing a lot of unnecessary lookups. I did not take the time to make a deep analysis of every expressions, but the `renderItem` (for account item rendering) was super slow, mostly because of `getLabelTextByAddress` and its lookups.

List of changes:
- Now use a `Map` instead of doing linear search when looking for the entropy source of every HD accounts.
- Added a reference to its `InternalAccount` counterpart on the `Account` (UI) type
- Added a `getLabelTextByInternalAccount` to avoid unnecessary account lookup

Here are some rough timings before the change with 70+ accounts:
```txt
 (NOBRIDGE) LOG  renderItem execution time (index= 0): 17.27ms
 (NOBRIDGE) LOG  renderItem execution time (index= 1): 22.62ms
 (NOBRIDGE) LOG  renderItem execution time (index= 2): 27.19ms
 (NOBRIDGE) LOG  renderItem execution time (index= 3): 23.45ms
 (NOBRIDGE) LOG  renderItem execution time (index= 4): 35.81ms
 (NOBRIDGE) LOG  renderItem execution time (index= 5): 39.84ms
 (NOBRIDGE) LOG  renderItem execution time (index= 6): 44.98ms
 (NOBRIDGE) LOG  renderItem execution time (index= 7): 49.31ms
 (NOBRIDGE) LOG  renderItem execution time (index= 8): 54.00ms
 (NOBRIDGE) LOG  renderItem execution time (index= 9): 57.89ms
 (NOBRIDGE) LOG  renderItem execution time (index=10): 69.77ms
 (NOBRIDGE) LOG  renderItem execution time (index=11): 3.29ms
 (NOBRIDGE) LOG  renderItem execution time (index=12): 34.81ms
 (NOBRIDGE) LOG  renderItem execution time (index=13): 27.28ms
 (NOBRIDGE) LOG  renderItem execution time (index=14): 35.06ms
 (NOBRIDGE) LOG  renderItem execution time (index=15): 35.38ms
 (NOBRIDGE) LOG  renderItem execution time (index=16): 37.87ms
 (NOBRIDGE) LOG  renderItem execution time (index=17): 38.12ms
 (NOBRIDGE) LOG  renderItem execution time (index=18): 83.29ms
 (NOBRIDGE) LOG  renderItem execution time (index=19): 93.72ms
 (NOBRIDGE) LOG  renderItem execution time (index=20): 97.36ms
 (NOBRIDGE) LOG  renderItem execution time (index=21): 102.34ms
 (NOBRIDGE) LOG  renderItem execution time (index=22): 106.34ms
 (NOBRIDGE) LOG  renderItem execution time (index=23): 109.99ms
 (NOBRIDGE) LOG  renderItem execution time (index=24): 114.87ms
 (NOBRIDGE) LOG  renderItem execution time (index=25): 126.37ms
 (NOBRIDGE) LOG  renderItem execution time (index=26): 131.16ms
 (NOBRIDGE) LOG  renderItem execution time (index=27): 135.40ms
 (NOBRIDGE) LOG  renderItem execution time (index=28): 139.91ms
 (NOBRIDGE) LOG  renderItem execution time (index=29): 144.31ms
 (NOBRIDGE) LOG  renderItem execution time (index=30): 148.63ms
 (NOBRIDGE) LOG  renderItem execution time (index=31): 153.94ms
 (NOBRIDGE) LOG  renderItem execution time (index=32): 157.74ms
 (NOBRIDGE) LOG  renderItem execution time (index=33): 161.69ms
 (NOBRIDGE) LOG  renderItem execution time (index=34): 175.08ms
 (NOBRIDGE) LOG  renderItem execution time (index=35): 178.71ms
 (NOBRIDGE) LOG  renderItem execution time (index=36): 183.05ms
 (NOBRIDGE) LOG  renderItem execution time (index=37): 187.29ms
 (NOBRIDGE) LOG  renderItem execution time (index=38): 191.40ms
 (NOBRIDGE) LOG  renderItem execution time (index=39): 197.23ms
 (NOBRIDGE) LOG  renderItem execution time (index=40): 200.39ms
 (NOBRIDGE) LOG  renderItem execution time (index=41): 204.13ms
 (NOBRIDGE) LOG  renderItem execution time (index=42): 210.40ms
 (NOBRIDGE) LOG  renderItem execution time (index=43): 221.05ms
 (NOBRIDGE) LOG  renderItem execution time (index=44): 226.08ms
 (NOBRIDGE) LOG  renderItem execution time (index=45): 222.74ms
 (NOBRIDGE) LOG  renderItem execution time (index=46): 234.90ms
 (NOBRIDGE) LOG  renderItem execution time (index=47): 240.98ms
 (NOBRIDGE) LOG  renderItem execution time (index=48): 243.44ms
 (NOBRIDGE) LOG  renderItem execution time (index=49): 247.17ms
 (NOBRIDGE) LOG  renderItem execution time (index=50): 252.63ms
 (NOBRIDGE) LOG  renderItem execution time (index=51): 258.43ms
 (NOBRIDGE) LOG  renderItem execution time (index=52): 262.03ms
 (NOBRIDGE) LOG  renderItem execution time (index=53): 273.43ms
 (NOBRIDGE) LOG  renderItem execution time (index=54): 269.31ms
 (NOBRIDGE) LOG  renderItem execution time (index=55): 281.84ms
 (NOBRIDGE) LOG  renderItem execution time (index=56): 287.93ms
 (NOBRIDGE) LOG  renderItem execution time (index=57): 291.56ms
 (NOBRIDGE) LOG  renderItem execution time (index=58): 303.75ms
 (NOBRIDGE) LOG  renderItem execution time (index=59): 307.33ms
 (NOBRIDGE) LOG  renderItem execution time (index=60): 305.86ms
 (NOBRIDGE) LOG  renderItem execution time (index=61): 310.80ms
 (NOBRIDGE) LOG  renderItem execution time (index=62): 321.12ms
 (NOBRIDGE) LOG  renderItem execution time (index=63): 318.82ms
 (NOBRIDGE) LOG  renderItem execution time (index=64): 331.08ms
 (NOBRIDGE) LOG  renderItem execution time (index=65): 333.92ms
 (NOBRIDGE) LOG  renderItem execution time (index=66): 338.62ms
 (NOBRIDGE) LOG  renderItem execution time (index=67): 350.39ms
 (NOBRIDGE) LOG  renderItem execution time (index=68): 356.76ms
 (NOBRIDGE) LOG  renderItem execution time (index=69): 352.59ms
 (NOBRIDGE) LOG  renderItem execution time (index=70): 356.64ms
 (NOBRIDGE) LOG  renderItem execution time (index=71): 368.87ms
 (NOBRIDGE) LOG  renderItem execution time (index=72): 368.52ms
 (NOBRIDGE) LOG  renderItem execution time (index=73): 377.89ms
 (NOBRIDGE) LOG  renderItem execution time (index=74): 382.32ms
```

> [!CAUTION]
> Something seemed to have an accumulative effect too, I did not dig up to understand what's going on, given that my new change do not have this behavior

- - -

Now with all the new change:
```txt
 (NOBRIDGE) LOG  renderItem execution time (index= 0): 2.82ms
 (NOBRIDGE) LOG  renderItem execution time (index= 1): 2.84ms
 (NOBRIDGE) LOG  renderItem execution time (index= 2): 2.83ms
 (NOBRIDGE) LOG  renderItem execution time (index= 3): 2.82ms
 (NOBRIDGE) LOG  renderItem execution time (index= 4): 2.85ms
 (NOBRIDGE) LOG  renderItem execution time (index= 5): 2.81ms
 (NOBRIDGE) LOG  renderItem execution time (index= 6): 2.81ms
 (NOBRIDGE) LOG  renderItem execution time (index= 7): 2.77ms
 (NOBRIDGE) LOG  renderItem execution time (index= 8): 2.84ms
 (NOBRIDGE) LOG  renderItem execution time (index= 9): 3.62ms
 (NOBRIDGE) LOG  renderItem execution time (index=10): 2.86ms
 (NOBRIDGE) LOG  renderItem execution time (index=11): 0.55ms
 (NOBRIDGE) LOG  renderItem execution time (index=12): 0.54ms
 (NOBRIDGE) LOG  renderItem execution time (index=13): 0.51ms
 (NOBRIDGE) LOG  renderItem execution time (index=14): 0.52ms
 (NOBRIDGE) LOG  renderItem execution time (index=15): 0.59ms
 (NOBRIDGE) LOG  renderItem execution time (index=16): 0.54ms
 (NOBRIDGE) LOG  renderItem execution time (index=17): 0.58ms
 (NOBRIDGE) LOG  renderItem execution time (index=18): 2.84ms
 (NOBRIDGE) LOG  renderItem execution time (index=19): 12.63ms
 (NOBRIDGE) LOG  renderItem execution time (index=20): 2.81ms
 (NOBRIDGE) LOG  renderItem execution time (index=21): 2.79ms
 (NOBRIDGE) LOG  renderItem execution time (index=22): 12.25ms
 (NOBRIDGE) LOG  renderItem execution time (index=23): 2.77ms
 (NOBRIDGE) LOG  renderItem execution time (index=24): 2.78ms
 (NOBRIDGE) LOG  renderItem execution time (index=25): 2.85ms
 (NOBRIDGE) LOG  renderItem execution time (index=26): 2.80ms
 (NOBRIDGE) LOG  renderItem execution time (index=27): 2.76ms
 (NOBRIDGE) LOG  renderItem execution time (index=28): 2.80ms
 (NOBRIDGE) LOG  renderItem execution time (index=29): 2.80ms
 (NOBRIDGE) LOG  renderItem execution time (index=30): 2.78ms
 (NOBRIDGE) LOG  renderItem execution time (index=31): 2.82ms
 (NOBRIDGE) LOG  renderItem execution time (index=32): 2.84ms
 (NOBRIDGE) LOG  renderItem execution time (index=33): 2.73ms
 (NOBRIDGE) LOG  renderItem execution time (index=34): 2.83ms
 (NOBRIDGE) LOG  renderItem execution time (index=35): 2.81ms
 (NOBRIDGE) LOG  renderItem execution time (index=36): 2.80ms
 (NOBRIDGE) LOG  renderItem execution time (index=37): 2.75ms
 (NOBRIDGE) LOG  renderItem execution time (index=38): 2.78ms
 (NOBRIDGE) LOG  renderItem execution time (index=39): 2.82ms
 (NOBRIDGE) LOG  renderItem execution time (index=40): 2.81ms
 (NOBRIDGE) LOG  renderItem execution time (index=41): 2.74ms
 (NOBRIDGE) LOG  renderItem execution time (index=42): 2.82ms
 (NOBRIDGE) LOG  renderItem execution time (index=43): 2.83ms
 (NOBRIDGE) LOG  renderItem execution time (index=44): 2.77ms
 (NOBRIDGE) LOG  renderItem execution time (index=45): 2.76ms
 (NOBRIDGE) LOG  renderItem execution time (index=46): 2.80ms
 (NOBRIDGE) LOG  renderItem execution time (index=47): 2.79ms
 (NOBRIDGE) LOG  renderItem execution time (index=48): 2.84ms
 (NOBRIDGE) LOG  renderItem execution time (index=49): 2.80ms
 (NOBRIDGE) LOG  renderItem execution time (index=50): 2.78ms
 (NOBRIDGE) LOG  renderItem execution time (index=51): 2.81ms
 (NOBRIDGE) LOG  renderItem execution time (index=52): 2.76ms
 (NOBRIDGE) LOG  renderItem execution time (index=53): 2.76ms
 (NOBRIDGE) LOG  renderItem execution time (index=54): 2.83ms
 (NOBRIDGE) LOG  renderItem execution time (index=55): 2.80ms
 (NOBRIDGE) LOG  renderItem execution time (index=56): 2.82ms
 (NOBRIDGE) LOG  renderItem execution time (index=57): 2.78ms
 (NOBRIDGE) LOG  renderItem execution time (index=58): 2.81ms
 (NOBRIDGE) LOG  renderItem execution time (index=59): 2.79ms
 (NOBRIDGE) LOG  renderItem execution time (index=60): 2.75ms
 (NOBRIDGE) LOG  renderItem execution time (index=61): 2.76ms
 (NOBRIDGE) LOG  renderItem execution time (index=62): 2.79ms
 (NOBRIDGE) LOG  renderItem execution time (index=63): 2.79ms
 (NOBRIDGE) LOG  renderItem execution time (index=64): 2.77ms
 (NOBRIDGE) LOG  renderItem execution time (index=65): 2.81ms
 (NOBRIDGE) LOG  renderItem execution time (index=66): 2.81ms
 (NOBRIDGE) LOG  renderItem execution time (index=67): 2.77ms
 (NOBRIDGE) LOG  renderItem execution time (index=68): 2.80ms
 (NOBRIDGE) LOG  renderItem execution time (index=69): 2.80ms
 (NOBRIDGE) LOG  renderItem execution time (index=70): 2.76ms
 (NOBRIDGE) LOG  renderItem execution time (index=71): 2.82ms
 (NOBRIDGE) LOG  renderItem execution time (index=72): 2.75ms
 (NOBRIDGE) LOG  renderItem execution time (index=73): 2.84ms
 (NOBRIDGE) LOG  renderItem execution time (index=74): 2.80ms
```

## **Related issues**

- https://github.com/MetaMask/metamask-mobile/issues/16345

## **Manual testing steps**

1. Onboard with multiple SRPs and a lot of accounts
2. Open the account list menu
  - It should be a bit faster and still work as usual

## **Screenshots/Recordings**

### **Before**

### **After**

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Contributor Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile Coding Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format if applicable
- [x] I’ve applied the right labels on the PR (see [labeling guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)). Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described in the ticket it closes and includes the necessary testing evidence such as recordings and or screenshots.