@ccharly 2025-06-28
> Is this for https://github.com/MetaMask/metamask-mobile/issues/16345? If so, might also want to tag it in the description to close the issue once this gets merged 🙂

It does not solve it entirely, but it is an improvement for it yes! I'll add it in the description 😄 thanks!